+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/AggregateTestNoIdReturns/DeleteAggregateTestNoIdReturn/DeleteAggregateTestNoIdReturnCommandHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using CleanArchitecture.TestApplication.Domain.Common.Exceptions;
 using CleanArchitecture.TestApplication.Domain.Repositories;
 using CleanArchitecture.TestApplication.Domain.Repositories.CRUD;
 using Intent.RoslynWeaver.Attributes;
@@ -26,6 +27,11 @@ namespace CleanArchitecture.TestApplication.Application.AggregateTestNoIdReturns
         public async Task<Unit> Handle(DeleteAggregateTestNoIdReturnCommand request, CancellationToken cancellationToken)
         {
             var existingAggregateTestNoIdReturn = await _aggregateTestNoIdReturnRepository.FindByIdAsync(request.Id, cancellationToken);
+
+            if (existingAggregateTestNoIdReturn is null)
+            {
+                throw new NotFoundException($"Could not find AggregateTestNoIdReturn {request.Id}");
+            }
             _aggregateTestNoIdReturnRepository.Remove(existingAggregateTestNoIdReturn);
             return Unit.Value;
         }
